requirements.txt
discord
pymongo
packaging
python-dotenv
sty